novel/requirements.txt
google-adk
python-dotenv
asyncio
typing 